Add asking for confirmation when deleting a server or a project

+++ ui/project_item.py
@@ -3,7 +3,7 @@
 This script contains ProjectItem.
 """
 
-from PyQt5.QtWidgets import QMenu, QAction, QTreeWidgetItem
+from PyQt5.QtWidgets import QMenu, QAction, QTreeWidgetItem, QMessageBox
 
 from qgis.core import QgsProject
 
@@ -73,8 +73,16 @@ class ProjectItem(QTreeWidgetItem):
         open_action.triggered.connect(open_)
 
         def delete():
-            self.qgis_server.delete_project(self.id)
-            self.server_item.removeChild(self)
+            confirm_dialog = QMessageBox(
+                QMessageBox.Question,
+                "Confirm project delete",
+                "Do you really want to delete this project?\n"
+                "It will be removed forever",
+                QMessageBox.Yes | QMessageBox.No,
+            )
+            if confirm_dialog.exec_() == QMessageBox.Yes:
+                self.qgis_server.delete_project(self.id)
+                self.server_item.removeChild(self)
         delete_action = QAction('Delete project')
         menu.addAction(delete_action)
         delete_action.triggered.connect(delete)
